agent-forge-state-generator/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
asyncpg==0.29.0
anthropic==0.7.7
pydantic==2.5.0
pydantic-settings==2.1.0
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
httpx==0.25.2
python-multipart==0.0.6
supabase==2.3.0
psycopg2-binary==2.9.9
